requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0
motor==3.3.2
pymongo==4.6.3
pydantic[email]==2.6.4
email-validator==2.1.1
python-dotenv==1.0.1
bcrypt==4.1.3
PyJWT==2.8.0
cloudinary==1.40.0
python-multipart==0.0.9
